gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
aiohttp==3.8.5
python-jose==3.3.0
python-multipart==0.0.6
jinja2==3.1.2
passlib==1.7.4
bcrypt==4.0.1
